chore: Migrate schedules.py to utils/convergence/, clean dead iterative/ code (#985)
- Move schedules.py to utils/convergence/schedules.py (iteration infrastructure)
- Update 2 import sites in coupling/fictitious_play.py, coupling/fixed_point_utils.py
- Delete dead prototype code: iterative/base.py, fixed_point.py, convergence.py,
  multi_population/, patterns/ (~500 lines)
- Leave deprecation shim at iterative/schedules.py for one version
- Clean alg/__init__.py: remove dead FixedPointSolver export, update docstring

Principle: alg/ = algorithms, utils/ = infrastructure.
Schedules are iteration infrastructure consumed by all paradigms.
