+++ mfgarchon/alg/numerical/coupling/fictitious_play.py
@@ -34,8 +34,8 @@ from typing import TYPE_CHECKING, Any
 
 import numpy as np
 
-# Import shared iteration utilities (Issue #630)
-from mfgarchon.alg.iterative.schedules import (
+# Import shared iteration utilities (Issue #630, migrated #985)
+from mfgarchon.utils.convergence.schedules import (
     LEARNING_RATE_SCHEDULES,
     get_schedule,
     harmonic_schedule,
